pyproject.toml
[project]
name = "yamllm-core"
version = "0.1.12"
description = "YAML-based LLM configuration and execution"
authors = [
    {name = "Daniel Halwell", email = "danielhalwell@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
]
dependencies = [
    "anthropic>=0.67.0",
    "bs4>=0.0.2",
    "duckduckgo-search>=7.5.5",
    "faiss-cpu>=1.10.0",
    "google-generativeai>=0.4.0",
    "httpx>=0.27.0",
    "matplotlib>=3.10.0",
    "mistralai>=1.7.1",
    "numpy>=2.2.3",
    "openai>=1.63.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=13.9.4",
    "ruff>=0.9.6",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.1.1",
    "isort>=5.13.2",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yamllm"]

[project.urls]
Homepage = "https://github.com/codehalwell/yamllm"
Documentation = "https://codehalwell.github.io/yamllm/"
Repository = "https://github.com/codehalwell/yamllm.git"

[project.scripts]
yamllm = "yamllm.cli:main"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
no_implicit_optional = true
strict_optional = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true
exclude = [
  "^tests/",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "azure-ai-inference>=1.0.0b1",
    "azure-identity>=1.15.0",
]
